Add files via upload
I’ve been learning data handling using Pandas in Python, and this is one of my practice projects. The dataset was downloaded from Kaggle and contains Dengue-related data from Bangladesh. In this task, I selected the 'AreaType' column as the feature and the 'Outcome' column as the target, then performed a train-test split for further analysis.